Move testable components' .h files next to source files
