+++ Common/app/imu_helper.cpp
@@ -14,7 +14,7 @@
 
 /********************************* Includes **********************************/
 #include "imu_helper.h"
-#include "dsp.h"
+#include "DSP/dsp.h"
