+++ Common/component/Dynamixel/AX12A.h
@@ -20,7 +20,7 @@
 
 
 /********************************* Includes **********************************/
-#include "Dynamixel.h"
+#include "Dynamixel/Dynamixel.h"
