+++ Common/component/Dynamixel/Dynamixel.h
@@ -22,7 +22,7 @@
 
 
 /********************************* Includes **********************************/
-#include "DaisyChain.h"
+#include "DaisyChain/DaisyChain.h"
