+++ Common/component/Dynamixel/MX28.h
@@ -21,7 +21,7 @@
 
 
 /********************************* Includes **********************************/
-#include "Dynamixel.h"
+#include "Dynamixel/Dynamixel.h"
