+++ Common/include/imu_helper.h
@@ -20,7 +20,7 @@
 
 
 /********************************* Includes **********************************/
-#include "MPU6050.h"
+#include "MPU6050/MPU6050.h"
