+++ Robot/.cproject
@@ -47,6 +47,7 @@
 									<listOptionValue builtIn="false" value="../../Common/mock"/>
 									<listOptionValue builtIn="false" value="../../Common/interface"/>
 									<listOptionValue builtIn="false" value="../../Common/hardware"/>
+									<listOptionValue builtIn="false" value="../../Common/component"/>
 								</option>
 								<option id="gnu.c.compiler.option.preprocessor.def.symbols.1620563188" name="Defined symbols (-D)" superClass="gnu.c.compiler.option.preprocessor.def.symbols" useByScannerDiscovery="false" valueType="definedSymbols">
 									<listOptionValue builtIn="false" value="__weak=__attribute__((weak))"/>
@@ -75,6 +76,7 @@
 									<listOptionValue builtIn="false" value="../../Common/mock"/>
 									<listOptionValue builtIn="false" value="../../Common/interface"/>
 									<listOptionValue builtIn="false" value="../../Common/hardware"/>
+									<listOptionValue builtIn="false" value="../../Common/component"/>
 								</option>
 								<option id="gnu.cpp.compiler.option.preprocessor.def.878794306" name="Defined symbols (-D)" superClass="gnu.cpp.compiler.option.preprocessor.def" useByScannerDiscovery="false" valueType="definedSymbols">
 									<listOptionValue builtIn="false" value="__weak=__attribute__((weak))"/>
@@ -187,6 +189,7 @@
 									<listOptionValue builtIn="false" value="../../Common/mock"/>
 									<listOptionValue builtIn="false" value="../../Common/interface"/>
 									<listOptionValue builtIn="false" value="../../Common/hardware"/>
+									<listOptionValue builtIn="false" value="../../Common/component"/>
 								</option>
 								<option id="gnu.c.compiler.option.preprocessor.def.symbols.674884498" name="Defined symbols (-D)" superClass="gnu.c.compiler.option.preprocessor.def.symbols" useByScannerDiscovery="false" valueType="definedSymbols">
 									<listOptionValue builtIn="false" value="__weak=__attribute__((weak))"/>
@@ -215,6 +218,7 @@
 									<listOptionValue builtIn="false" value="../../Common/mock"/>
 									<listOptionValue builtIn="false" value="../../Common/interface"/>
 									<listOptionValue builtIn="false" value="../../Common/hardware"/>
+									<listOptionValue builtIn="false" value="../../Common/component"/>
 								</option>
 								<option id="gnu.cpp.compiler.option.preprocessor.def.1119572749" name="Defined symbols (-D)" superClass="gnu.cpp.compiler.option.preprocessor.def" useByScannerDiscovery="false" valueType="definedSymbols">
 									<listOptionValue builtIn="false" value="__weak=__attribute__((weak))"/>
