+++ RobotTest/.cproject
@@ -53,6 +53,7 @@
 									<listOptionValue builtIn="false" value="../../Common/interface"/>
 									<listOptionValue builtIn="false" value="../../Common/mock"/>
 									<listOptionValue builtIn="false" value="../../Common/hardware"/>
+									<listOptionValue builtIn="false" value="../../Common/component"/>
 								</option>
 								<option id="gnu.cpp.compiler.option.preprocessor.def.1235015124" name="Defined symbols (-D)" superClass="gnu.cpp.compiler.option.preprocessor.def" useByScannerDiscovery="false" valueType="definedSymbols">
 									<listOptionValue builtIn="false" value="STM32F767xx"/>
@@ -89,6 +90,7 @@
 									<listOptionValue builtIn="false" value="../../Common/interface"/>
 									<listOptionValue builtIn="false" value="../../Common/mock"/>
 									<listOptionValue builtIn="false" value="../../Common/hardware"/>
+									<listOptionValue builtIn="false" value="../../Common/component"/>
 								</option>
 								<option id="gnu.c.compiler.option.preprocessor.def.symbols.1393467156" name="Defined symbols (-D)" superClass="gnu.c.compiler.option.preprocessor.def.symbols" useByScannerDiscovery="false" valueType="definedSymbols">
 									<listOptionValue builtIn="false" value="STM32F767xx"/>
@@ -167,6 +169,7 @@
 									<listOptionValue builtIn="false" value="../../Common/interface"/>
 									<listOptionValue builtIn="false" value="../../Common/mock"/>
 									<listOptionValue builtIn="false" value="../../Common/hardware"/>
+									<listOptionValue builtIn="false" value="../../Common/component"/>
 								</option>
 								<option id="gnu.cpp.compiler.option.preprocessor.def.245366192" name="Defined symbols (-D)" superClass="gnu.cpp.compiler.option.preprocessor.def" useByScannerDiscovery="false" valueType="definedSymbols">
 									<listOptionValue builtIn="false" value="STM32F446xx"/>
@@ -193,6 +196,7 @@
 									<listOptionValue builtIn="false" value="../../Common/interface"/>
 									<listOptionValue builtIn="false" value="../../Common/mock"/>
 									<listOptionValue builtIn="false" value="../../Common/hardware"/>
+									<listOptionValue builtIn="false" value="../../Common/component"/>
 								</option>
 								<option id="gnu.c.compiler.option.preprocessor.def.symbols.302701397" name="Defined symbols (-D)" superClass="gnu.c.compiler.option.preprocessor.def.symbols" useByScannerDiscovery="false" valueType="definedSymbols">
 									<listOptionValue builtIn="false" value="STM32F446xx"/>
